+++ common/ktor/src/main/kotlin/no/nav/mulighetsrommet/ktor/extensions/ApplicationCall.kt
@@ -1,35 +1,22 @@
 package no.nav.mulighetsrommet.ktor.extensions
 
 import io.ktor.http.*
+import io.ktor.http.auth.*
 import io.ktor.server.application.*
+import io.ktor.server.auth.*
 import no.nav.mulighetsrommet.ktor.exception.StatusException
 
 /**
- * Utility extension on [ApplicationCall] to get a path parameter by [name].
+ * Utility extension on [ApplicationCall] to get the bearer token from the request.
  *
- * A [StatusException] will be thrown if the parameter is null or empty.
+ * A [StatusException] will be thrown if the bearer token is missing.
  */
-fun ApplicationCall.getNonEmptyPathParameter(name: String): String {
-    val parameter = parameters[name]
+fun ApplicationCall.getAccessToken(): String {
+    val header = request.parseAuthorizationHeader()
 
-    if (parameter.isNullOrEmpty()) {
-        throw StatusException(HttpStatusCode.BadRequest, "Path parameter $name is missing")
+    if (header is HttpAuthHeader.Single && header.authScheme.lowercase() in listOf("bearer")) {
+        return header.blob
     }
 
-    return parameter
-}
-
-/**
- * Utility extension on [ApplicationCall] to get a query parameter by [name].
- *
- * A [StatusException] will be thrown if the parameter is null or empty.
- */
-fun ApplicationCall.getNonEmptyQueryParameter(name: String): String {
-    val parameter = request.queryParameters[name]
-
-    if (parameter.isNullOrEmpty()) {
-        throw StatusException(HttpStatusCode.BadRequest, "Query parameter $name is missing")
-    }
-
-    return parameter
+    throw StatusException(HttpStatusCode.Unauthorized, "Mangler accesstoken - Ingen tilgang")
 }
